Add Windows (MSVC) build support and fix engine warnings
Windows was declared as an ffiPlugin target but had never actually been
compiled. Three untested-on-Windows spots in the native CMake, plus three
MSVC warnings, are fixed here (all behavior-preserving, so Apple/Android
builds are unaffected):

- CMake: add the iPlug2/WDL include dir so win32_utf8.h resolves. It is only
  pulled in under `#ifdef _WIN32`, so Apple/Android never needed it.
- CMake: add an MSVC anchor translation unit (wajuce_msvc_anchor.cpp). link.exe
  will not emit a DLL for the otherwise header-only `wajuce` shared target
  (Apple's -force_load / Linux's --whole-archive tolerate it, MSVC does not);
  the engine's dllexport'd wajuce_* symbols re-export via /WHOLEARCHIVE.
- C4244: cast the std::exp() result to float in ParamAutomation SetTarget so
  the RHS stays float (value is float).
- C4190: wrap the static getBridgeState helper in `extern "C++"` — it sat
  inside the FFI `extern "C"` block but returns a shared_ptr.
- C4018: make the channels * bytesPerSample product uint32_t to match pcmBytes
  in the WAV decoder bounds check.

+++ native/engine/Source/ParamAutomation.h
@@ -217,8 +217,9 @@ private:
       case AutomationEventType::SetTarget:
         if (event.timeConstant > 0.0f) {
           value = event.value + (value - event.value) *
-                                    std::exp(-(time - event.time) /
-                                             event.timeConstant);
+                                    static_cast<float>(
+                                        std::exp(-(time - event.time) /
+                                                 event.timeConstant));
         }
         break;
       case AutomationEventType::LinearRamp:
